+++ Reddit/Easy/integer_complexity_1/src/lib.rs
@@ -1,0 +1,7 @@
+
+pub fn stub() {
+    println!("hello there");
+}
+
+#[cfg(test)]
+mod tests;