+++ Reddit/Easy/integer_complexity_1/src/main.rs
@@ -1,0 +1,7 @@
+extern crate integer_complexity_1;
+
+use integer_complexity_1 as complexity;
+
+fn main() {
+    complexity::stub();
+}