+++ Reddit/Easy/integer_complexity_1/src/tests.rs
@@ -1,0 +1,5 @@
+
+#[test]
+fn stub_test() {
+    assert_eq!(1, 1);
+}